feat: added social authentication, core upgrade improvement and cache cleanup

+++ database/migrations/2023_01_01_000003_create_posts_table.php
@@ -14,6 +14,10 @@ return new class () extends Migration {
      */
     public function up(): void
     {
+        if (Schema::hasTable('posts')) {
+            return;
+        }
+
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
